Use `getBatched` in Firefox 52+.
The Python and Go servers behave differently when fetching records
without a limit: Python assumes "no limit" and returns everything,
while Go uses the default limit (1k). This meant the validator only
fetched a maximum of 1k records when talking to a Go server, and
reported all other records as "server missing".

In mozilla-services/go-syncstorage#166, @rfk suggested using download
batching instead of relying on the absence of a limit.