Add an option to enable Browser Console logging.
In bug 1617503, a user was having issues with a wedged Places
connection. I was curious if that was caused by a stuck merge runnable,
so I asked the user to enable trace logging. However, Desktop doesn't
flush log files to disk (via `resetFileLog`) until the sync finishes
or there's an error—see `policies.js` for those cases. This means
log files aren't always helpful for seeing stuck syncs.

We do provide an option for `dump` logging, but that requires
launching Firefox from the command line, which is tricky if you're
not used to doing that, and also includes lots of other warnings and
log output. The Browser Console is much easier to access, and also
gives us live logs. We already have a pref controlling that, and this
patch just surfaces it in the UI.